Flask Application/requirements.txt
Flask==3.0.3
numpy==2.0.2
tensorflow==2.18.0
keras==3.6.0
opencv-python==4.10.0.84
h5py==3.12.1
protobuf==5.28.3
requests==2.32.3
Pillow==11.0.0
